+++ flower-integration-github/src/flower/tracker/github/task.clj
@@ -16,13 +16,13 @@
 ;; Public definitions
 ;;
 
-(defrecord GithubTrackerTask [tracker task-id task-title task-type task-state task-tags]
+(defrecord GithubTrackerTask [tracker task-id task-title task-type task-state task-tags task-description]
   proto/TrackerTaskProto
   (get-tracker [tracker-task] tracker)
   (get-task-id [tracker-task] task-id)
   (get-state [tracker-task] task-state)
   (get-type [tracker-task] task-type)
-  (update! [tracker-task] (private-set-github-workitem! tracker-task)))
+  (upsert! [tracker-task] (private-set-github-workitem! tracker-task)))
 
 
 (macros/public-definition get-github-workitems cached)
@@ -45,7 +45,8 @@
           :task-state (.getState %)
           :task-tags (map (fn [label]
                             (.toString label))
-                          (.getLabels %))})
+                          (.getLabels %))
+          :task-description (.getBody %)})
        (if (empty? task-ids)
          (common/get-github-workitems-inner tracker)
          (common/get-github-workitems-inner tracker task-ids))))
@@ -62,8 +63,8 @@
   (let [tracker (proto/get-tracker tracker-task)
         task-id (proto/get-task-id tracker-task)
         old-workitem (first (proto/get-tasks tracker [task-id]))
-        fields (second (data/diff old-workitem tracker-task))]
-    (common/set-github-workitem-inner! tracker task-id fields)
+        fields (second (data/diff old-workitem tracker-task))
+        new-task-id (.getNumber (common/set-github-workitem-inner! tracker task-id fields))]
     (common/get-github-workitems-inner-clear-cache!)
     (get-github-workitems-clear-cache!)
-    tracker-task))
+    (first (proto/get-tasks tracker [new-task-id]))))