+++ flower-integration-gitlab/src/flower/tracker/gitlab/task.clj
@@ -16,13 +16,13 @@
 ;; Public definitions
 ;;
 
-(defrecord GitlabTrackerTask [tracker task-id task-title task-type task-state task-tags]
+(defrecord GitlabTrackerTask [tracker task-id task-title task-type task-state task-tags task-description]
   proto/TrackerTaskProto
   (get-tracker [tracker-task] tracker)
   (get-task-id [tracker-task] task-id)
   (get-state [tracker-task] task-state)
   (get-type [tracker-task] task-type)
-  (update! [tracker-task] (private-set-gitlab-workitem! tracker-task)))
+  (upsert! [tracker-task] (private-set-gitlab-workitem! tracker-task)))
 
 
 (macros/public-definition get-gitlab-workitems cached)
@@ -43,7 +43,8 @@
                              (.getUsername assignee)
                              nil))
           :task-state (.getState %)
-          :task-tags (seq (.getLabels %))})
+          :task-tags (seq (.getLabels %))
+          :task-description (.getDescription %)})
        (if (empty? task-ids)
          (common/get-gitlab-workitems-inner tracker)
          (common/get-gitlab-workitems-inner tracker task-ids))))
@@ -60,8 +61,8 @@
   (let [tracker (proto/get-tracker tracker-task)
         task-id (proto/get-task-id tracker-task)
         old-workitem (first (proto/get-tasks tracker [task-id]))
-        fields (second (data/diff old-workitem tracker-task))]
-    (common/set-gitlab-workitem-inner! tracker task-id fields)
+        fields (second (data/diff old-workitem tracker-task))
+        new-task-id (.getIid (common/set-gitlab-workitem-inner! tracker task-id fields))]
     (common/get-gitlab-workitems-inner-clear-cache!)
     (get-gitlab-workitems-clear-cache!)
-    tracker-task))
+    (first (proto/get-tasks tracker [new-task-id]))))