+++ flower-integration-tfs/src/flower/tracker/tfs/task.clj
@@ -17,13 +17,13 @@
 ;; Public definitions
 ;;
 
-(defrecord TFSTrackerTask [tracker task-id task-title task-type task-state task-tags]
+(defrecord TFSTrackerTask [tracker task-id task-title task-type task-state task-tags task-description]
   proto/TrackerTaskProto
   (get-tracker [tracker-task] tracker)
   (get-task-id [tracker-task] task-id)
   (get-state [tracker-task] task-state)
   (get-type [tracker-task] task-type)
-  (update! [tracker-task] (private-set-tfs-workitem! tracker-task)))
+  (upsert! [tracker-task] (private-set-tfs-workitem! tracker-task)))
 
 
 (macros/public-definition get-tfs-workitems cached)
@@ -43,7 +43,8 @@
             :task-assignee (get fields :System.AssignedTo)
             :task-state (get fields :System.State)
             :task-tags (filter (complement empty?)
-                               (string/split (get fields :System.Tags "") #"; "))}))
+                               (string/split (get fields :System.Tags "") #"; "))
+            :task-description (get fields :System.Description)}))
        (if (string? query)
          (common/get-tfs-query-inner tracker query)
          (common/get-tfs-workitems-inner tracker query))))
@@ -62,17 +63,19 @@
   (let [tracker (proto/get-tracker tracker-task)
         task-id (proto/get-task-id tracker-task)
         old-workitem (first (proto/get-tasks tracker [task-id]))
-        diff (second (data/diff old-workitem tracker-task))
+        diff (into {} (filter second (second (data/diff old-workitem tracker-task))))
         fields (set/rename-keys diff {:task-title :System.Title
                                       :task-type :System.WorkItemType
                                       :task-assignee :System.AssignedTo
                                       :task-state :System.State
-                                      :task-tags :System.Tags})
-        fields-with-tags (if (contains? fields :System.Tags)
-                           (assoc fields :System.Tags (string/join "; "
-                                                                   (get fields :System.Tags [])))
-                           fields)]
-    (common/set-tfs-workitem-inner! tracker task-id fields-with-tags)
+                                      :task-tags :System.Tags
+                                      :task-description :System.Description})
+        fields-with-tags (dissoc (if (contains? fields :System.Tags)
+                                   (assoc fields :System.Tags (string/join "; "
+                                                                           (get fields :System.Tags [])))
+                                   fields)
+                                 :tracker)
+        new-task-id (get (common/set-tfs-workitem-inner! tracker task-id fields-with-tags) :id)]
     (common/get-tfs-workitems-inner-clear-cache!)
     (get-tfs-workitems-clear-cache!)
-    tracker-task))
+    (first (proto/get-tasks tracker [new-task-id]))))